requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Authentication & Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Stripe
stripe>=5.5.0

# Existing dependencies from legal_rag_chatbot.py
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
qdrant-client>=1.6.9
colorama>=0.4.6
torch>=2.3.1